Add high-level Java API
The API is inspired by the standard library `Pattern` class, but tries
to be a little bit more direct (the intermediate `Matcher` is
unnecessary) - one big batch of allocations during the creation step
suffices.

The `DfaPattern.compile` is where the next bit of work is headed:
dynamically creating bytecode for the DFAs.