Add types for managing ranges/sets of code points
The main limitation of the current regex support is that it cannot
efficiently handle ranges of characters. This means that even simple
stuff like `\w` or `.` would need to be expanded into a lot of cases.

This commit adds support for `CodeUnitRange` to represent a range of
16-bit code units and `CodeUnitSet` to represent a set of 16-bit code
units. The latter supports all of the usual boolean operations.