Reduce states in M2 & add initial JMH benchmarks
M1 and M2 now associate character transitions with _multiple_
characters. This means that character classes don't need to be desugared
as nested unions, which in turn means that there isn't an explosion of
states in M2 and M3.

The initial benchmarks are somewhat promising: the DFA approach is
beating out the default Java engine (checking is _much_ faster, capture
is faster).