Add ScalaCheck parsing tests
There is a new `ReGenerator` trait that carries an `Arbitrary[Re]`
instance. This instance generates ASTs for valid regular expressions.
Combined with a new `RegexPrinter` visitor, it can be used to
synthesize arbitrary regular expressions and their source forms.

Right now, this is just used to assert that pretty-printing and
re-parsing round-trips the regex AST. In the future, this will be used
to automatically generate patterns, matching inputs, and failing inputs.

A couple parsing bugs or missing features were uncovered along the way:

  - Fix precendence of intersection/union (union binds more tightly)
  - Support the `.`, `\h`, `\H`, `\v`, `\V` character classes
  - Support the `\b`, `\B`, `\A`, `\Z`, `\z` boundary patterns
  - Support control character escapes `\cX`
  - Support octal characters escapes `\0123`