Optimization: TDFA minimization, use `tableswitch`
TDFA minimization follows Hopcroft's algorithm, modified to take into
account the tags on the TDFA. Depending on the regular expression, this
potentially reduces the number of states by quite a bit!

In code generation, add a very small optimization to use `tableswitch`
instead of `lookupswitch` if the values being tested are consecutive.
This isn't about efficiency of the generate code (they should be
equivalent), but rather about the size of the emitted code.