Optimization: detect fixed tags
Detect group markers that are at fixed distances from other tags. For
those cases, avoid tracking the secondary tags - just compute them at
the end of the match. Apparently this often has the effect of reducing
the number of registers and commands.

This was implemented based on the description in the "Fixed tags"
section of the 2017 Trofimovich TDFA paper.