Generate TDFAs for matching prefixes of inputs
This adds support for `lookingAt` functionality: the TDFA will accept
and return the prefix of the input which matched the regular expression.
This is the last theoretical piece needed for supporting `find` too
(although that's not done in this commit).

The TNFA is unaffected, but the TDFA now has an additional "prefix"
mode.

  - execution: under the prefix mode, encountering a code unit which has
    no transition only means failure if the state doesn't have a final
    transition. If it does, we use the final transition and then
    succeed.

  - compilation: all states after an accepting state are marked as
    "accepting" and assignment to group markers occurs on all outgoing
    transitions from a final state (since we may need those group
    markers backed up).

One key insight: the semantics we aim for with `lookingAt` are those of
Java, so we're not always trying to get the longest matching prefix (it
really depends on the quantifiers). This is achievable by skipping row
configurations which have a lower priority that a row corresponding to
the final TNFA state. The new `RegexLookingAtSpec` suite has some
examples of this.