Generate TNFAs for lazily skipping prefixes of inputs
This adds support for `find` functionality: the TNFA can be prefixed by
something similar to `.*?` so that any number of codepoints can be
skipped before the actual pattern is reached.

After the TNFA is built, nothing in the rest of the compilation pipeline
changes. Added a `results` method for easily listing out all
non-overlapping found from the regular expression.