Optimization: better fixed group marker tracking
The new `GroupMarkers` class handles building up and managing the
information about which fixity equivalence classes exist and which
of them are anchored to the start/end of the input.

One of the outcomes is that `captureMatch` on a pattern with no explicit
groups (just the one implicit first group) doesn't need any tag
commands. The start and end of the group are fixed to the start and end
of the input. As a result, we can remove `checkMatch` from the API and
from the code generation: the same DFA (with no commands) will anyways
be generated by using `captureMatch`.

The `DfaPattern` API is getting closer the JDK one: the three matching
modes are now exactly the three methods + TDFAs.