API revamp to use a mutable `DfaMatcher`
This (by design) should be quite close the Java API. The idea behind the
new mutable `DfaMatcher` object is to enable multiple successive
matching calls on the same input and to re-use the groups array across
multiple matches (instead of needing to reallocate it every time).

This does seem to slow down things on the order of single-digit
percentages.