Add `tester` driver for validating against JDK tests
This adds a tester tool that runs test cases in files matching the
syntax and structure of the OpenJDK tests found in
`test/jdk/java/util/regex/*.txt`. For example, assuming you've
downloaded `TestCases.txt`:

    sbt 'tester/run TestCases.txt'

About 20% of tests are failing at the moment, although they seem mostly
due to features and flags that are still missing.