Bugfix: avoid coalescing away group markers
This works around a nasty bug which manifested non-deterministically as
groups in find DFAs sometimes "missing" (the interpreted variant would
throw a helpful error like "S0 should be unavoidable").

The root cause turned out to be that sometimes a group marker would get
coalesced into a temporary register, and then the temporary register
wouldn't be marked as `liveOut` in the next iteration, leading it to be
eliminated in DSE. The current solution is really still a hack: we just
skip any coalesce that would get rid of a group marker.

The entire story around choice and ordering of coalescings needs to be
revisited - it is clear that the order in which variables get coalesced
has a massive impact on DFA minimization, so perhaps the steps could be
interleaved.